backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
hiredis==2.3.2

# Neo4j
neo4j==5.15.0

# LangChain & LangGraph
langchain==0.1.4
langchain-openai==0.0.5
langchain-anthropic==0.1.1
langchain-community==0.0.16
langgraph==0.0.20
langsmith>=0.0.83

# AI/ML
openai==1.10.0
anthropic>=0.17.0
tiktoken==0.5.2
sentence-transformers==2.3.1

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2024.1

# Monitoring & Logging
prometheus-client==0.19.0
structlog==24.1.0

# Task Queue
celery==5.3.6
flower==2.0.1

# Email
python-email-validator==2.1.0

# File handling
python-magic==0.4.27
pillow==10.2.0

# Excel/CSV
openpyxl==3.1.2
pandas==2.2.0

# Odoo Integration
xmlrpc==1.0.1

# Rate Limiting
slowapi==0.1.9

# CORS
fastapi-cors==0.0.6
